clean up UDF
